+++ embedded-graphics/src/pixelcolor/mod.rs
@@ -53,21 +53,21 @@
 //!
 //! egrectangle!(
 //!     top_left = (0, 0),
-//!     bottom_right = (100, 100),
+//!     size = (100, 100),
 //!     style = primitive_style!(fill_color = EpdColor::White)
 //! )
 //! .draw(&mut display)?;
 //!
 //! egrectangle!(
 //!     top_left = (100, 0),
-//!     bottom_right = (200, 100),
+//!     size = (100, 100),
 //!     style = primitive_style!(fill_color = EpdColor::Black)
 //! )
 //! .draw(&mut display)?;
 //!
 //! egrectangle!(
 //!     top_left = (200, 0),
-//!     bottom_right = (300, 100),
+//!     size = (100, 100),
 //!     style = primitive_style!(fill_color = EpdColor::Red)
 //! )
 //! .draw(&mut display)?;